RENAME : org.eclipse.linuxtools.systemtap.localgui.tests =>org.eclipse.linuxtools.systemtap.local.tests
THis test suite now contains only tests that depend solely on core and launch. Tests that refer to .callgraph have been moved to another test suite, and tests that contained a mixture of both graphing and non-graphing tests have been split into graphing and non-graphing components to be placed in their respective locations. Error in SystemTapDisplayTest and SystemTapGraphTest and StapGraphParserTest is intentional - SVN is refusing the changes otherwise. Will delete in next commit.