Plethora of changes to make SystemTapTextView work properly for non-realtime functions
Create an update method in SystemTapView -- this method is called by the view when it is initialized by a non-realtime parser.

Add finalData object to SystemTapParser -- the user must define a setFinalData method which is called immediately after executeParsing, and the getFinalData object is used by view.update to update the view. For real-time functionality, leave these functions blank.

Make the appropriate changes to GraphUIJob to call view.update, change SystemTapLaunchDelegate to properly pass a submonitor into the parser.