Enable source features and bundles generation.
Source feature generation is activated if there is sourceTemplateFeature directory in the feature. For the sake of git not caring about empty directories I put empty feature.properties in these directories.
Also a p2 plugin instruction is added to the poms in order to get the generated source features visible for p2.
New category for the sources is added to the repository category.xml and source features are added to it.