Fix format in Systemtap help pages
Fix formatting of SystemTap Help pages which are not yet in wikis, and
convert them to XHTML.  This patch can be considered a clean-up step to
convert the pages to Wiki format.

* Hooked up all of the HTML files to a per-project CSS file that was
originally taken from the CDT project.  This gives the Help files
consistent formatting with the rest of the Help pages in Eclipse.

* To make it easier to find errors in the formatting metadata, I've
switched the tags to use XHTML which is more strict.  This required
fixing a lot of incorrectly coded, and in a few cases, broken HTML.

* In the process of doing this work, the Eclipse web page editor pointed
out some spelling errors, which I have fixed.

Signed-off-by: Corey Ashford <cjashfor@linux.vnet.ibm.com>