tmf: Expose a method to retrieve an attribute's name
In the past, the only way to retrieve quarks was to pass its
exact name, so it implied we knew the name in the first place...

But with wildcard methods like getQuarks(), we won't necessarily
know the name of the quarks we receive. This new method will
allow the user to do a reverse lookup on the name.

Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>