lttng: Don't update previous data wrt sched_switch trace event
Some event types in the event handler update past data
(to correctly identify "unknown" threads for example),
but this should not be done in the case of sched_switch.

The problem is that sched_switch is always called after
a sched_exit, and since the sched_exit's handler deletes
all information related to this process, we do not want
the following sched_switch to restore the Exec_name for
example.

Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>