lttng: Always create the sub-attributes of TID nodes
This makes the history slightly bigger, but it avoids having
to catch all sort of exceptions in the views, for the cases
where the expected sub-attributes never get populated. Now
they will simply contain a "null" value for the whole duration.

Also relaxed the test cases a bit, changed it so we won't have
to update them every time we make a single change to the state
provider.

Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>