tmf: Make HistoryBuilder use a TMF request for its events
Initially, the IStateChangeInput object would have to open
the trace itself, and merely send state changes. Rework this
interface so that we now have to pass it the events. This
allows HistoryBuilder to hook into TMF's event requests, and
benefit from the coalescing.

This will be tweaked a bit more soon, when we put all indexing
requests together.

Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>