lttng: Make sure we assign a value to all created quarks
Rework the fix in f3013ba. If a quark gets created but is never
ever used over the span of the whole history, it ends up not
having any value whatsover, not even null. At query time, it
would query the empty space in the HT nodes for its value,
returning 0 most of the time.

Now we manually assign a value to all Thread sub-attributes,
depending on the event type.

Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>