tmf: Internalize the state history storage backend
Initially, the design was to allow both state providers
(inputs) and storage backends to be extended from external
sources. The state input obviously has to be public,
since the support for specific trace types is implemented
in separate plugings. However, there is no immediate need
to expose the storage backend API, since the only implementation
resides in TMF.

This could easily be revisited in the future if we want to
support external state storage backends. For now, it'll
simplify TMF's public API.

This is candidate for the 1.0 branch, since it's only refactoring.

Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>