Systemtap: Showing warning in ViewPart when kernel source is not defined
Before this commit the plugin was asking the path for the user using a
File Dialog. It was not clear what was being asked and this will be
harder to implement for remote projects.
So it was replaced by a warning message in the Kernel Source Browser
ViewPart with instructions in how to set the Kernel Source Path in
SystemTap preferences.