Re-write of perf plug-in tests.
Tests were re-written to be less machine dependent. Pre-generated data
files are under resource/*-event. In the tests assertions are made about
these specific files, allowing for consistent results across machines.

The parsing of the perf record (PerfCore.java: Record(...)) is now
handled in isolation to allow for other types of input streams to be
passed in (e.g. text files). This was also done for the annotation to
make an explicit separation between parsing a report and an annotation.

Tests involving unresolved symbols were removed, as the feature is not
fully implemented (commented out,in PerfCore.java: getReportString(...)).

checkChildrenStructure was modified to not accept elements with no
children, as an even element can have no children.

checkChildrenPercentages was modified to allow for acceptable percentage
ranges to compensate for truncation errors.