Initial contribution of Profiling Unification work.
org.eclipse.linuxtools.profiling.launch.launchProvider is an extension
point created to provide launch shortcuts for various profiling types,
all of which must extend ProfileLaunchShortcut, and specify a profiling
type that they provide.

org.eclipse.linuxtools.profiling.snapshot is a plugin (it will be one
among a few others) that provides a basic launch shortcut. When
launched, it will find a launch shortcut provided by the launchProvider
extension point which has a type it supports, and launch it.