tmf: Close the state history files before deleting them
When a state history construction gets cancelled, we delete the
partial files, so we are not stuck with an incomplete file that
looks exactly like a complete and valid one.

However, we should close the file descriptors before attempting the
deletion. On Linux, lsof would still report the descriptors to the
deleted files after we stopped using them. On Windows it was even
worse, the deletion would fail because the file lock was still taken.

Change-Id: I29d0bf39621bdf6713bb2c3fe908e16f8d4f2bcc
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>