lttng: Handle the statedump events
Back when the state provider was first developed, LTTng did
not support the statedump (dummy events that show the current
state of all the running processes).

This support was added to upstream LTTng since then, so the
TMF viewer should make good use of this information!

Change-Id: I194fa8f31051b5755ed35e1594d33e47dcceec7a
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/7032
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
IP-Clean: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>