lttng: Fix off-by-one in the check for new attributes
In some places, we want to check if a get*AndAdd() call just
created an attribute, or re-used an existing one. The correct
way to check for this is if (quark == ss.getNbAttributes() - 1).
The -1 was missing.

Since this test comes relatively often, also add a specific method
to the interface, so it makes the code clearer.

Change-Id: Idc3f9a508cd670914236bce603e1d5202787a46c
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/7103
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
IP-Clean: Matthew Khouzam <matthew.khouzam@ericsson.com>