tmf: Generalization of the statistics view
The creation of the statistics view is now based on the type of the
trace selected and his plugin definition. Thus, it is easier to provide
a new implementation of the statistics for a different kind of trace.

Change-Id: If105007a5f349565c14cdf6e5e084a4ab5d227b3
Signed-off-by: Mathieu Denis <mathieu.denis@polymtl.ca>
Reviewed-on: https://git.eclipse.org/r/7041
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
IP-Clean: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-by: Bernd Hufmann <bhufmann@gmail.com>
IP-Clean: Bernd Hufmann <bhufmann@gmail.com>