Fix a sporadic distribution problem in the TMF coalesced requests
When a data/event provider receives a set of requests during the
processing of a signal, it will do its best to coalesce these requests
in order to minimize the number of physical accesses that need to be
performed on the back-end traces.

The coalesced request, itself being a request, also has a data type. By
default, its data type is determined by the first sub-request coalesced
by the data/event provider. Afterwards, when it receives data from the
back-end, it dispatches the data/events to the appropriate sub-request.

This works fine for homogeneous experiments (same trace/event types).
However, when dealing with heterogeneous traces or even with different
event types during a given session, the coalesced event type, being
determined by the first sub-request, can in fact hide the data/events
for sub-requests whose data type is not an extension of the base type.

This fix forces the data type of the coalesced request to ITmfEvent
therefore ensuring that the data meant for the sub-requests is not
inadvertently filtered out.


Change-Id: I2862f7ca30d38c968968dd6cd41f6cc04bbcd9a4
Signed-off-by: Francois Chouinard <fchouinard@gmail.com>
Reviewed-on: https://git.eclipse.org/r/7171
Reviewed-by: Bernd Hufmann <bhufmann@gmail.com>
IP-Clean: Bernd Hufmann <bhufmann@gmail.com>
Tested-by: Bernd Hufmann <bhufmann@gmail.com>