tmf: Use tabs in statistics view for each traces
When an experiment/trace is selected, the statistics view creates a
global viewer and a specific viewer for each trace contained in the
experiment. These viewers are placed in tabs and shown in the
statistics view. Also, a factory is now used to create a viewer based
on his definition in the plug-in.xml from the trace type. An
extension point element has been added to the trace type extension
point to provide custom statistics viewer.

Fixes bug 387217 on bugs.eclipse.org

Signed-off-by: Mathieu Denis <mathieu.denis@polymtl.ca>
Change-Id: I8d380e8c28cba855b5361069200e50ab0cbca7f3
Reviewed-on: https://git.eclipse.org/r/7042
Reviewed-by: Bernd Hufmann <bhufmann@gmail.com>
IP-Clean: Bernd Hufmann <bhufmann@gmail.com>
Tested-by: Bernd Hufmann <bhufmann@gmail.com>