tmf: Correctly use null values for empty stacks
When removing the last element of a stack attribute, the
popAttribute() method would leave the stack-attribute (which
represents the depth of the stack) with a state value of 0.

We expect a null value for empty stacks, and the integer
representation of a null value is -1. This would make it
impossible to re-use an attribute stack after it was completely
emptied once.

This patch fixes this by correctly assigning a null value when
emptying a stack.

Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Change-Id: If2d4edc8a0bee55f007a4dcbffefe44f513f567d
Reviewed-on: https://git.eclipse.org/r/7288
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
IP-Clean: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>