Revert "Profiling type launch shortcuts should call provider's launch shortcut."
This reverts commit 24c6de80d252339b51b2ab11a093d34d91f1d91f, which separated
launching from a shortcut and launching from a delegate. This functionality
was not correct as it did not allow to change the underlying tool which was
providing the delegate and shortcut classes. In addition, the configuration
passed on to the delagate is updated to include plug-in specfic launch
configuration attributes.

Change-Id: I34131c7e6ddd26c141eabc7a89030795e0f4a38f
Reviewed-on: https://git.eclipse.org/r/7627
Tested-by: Hudson CI
Reviewed-by: Roland Grunberg <rgrunber@redhat.com>
IP-Clean: Roland Grunberg <rgrunber@redhat.com>
Tested-by: Roland Grunberg <rgrunber@redhat.com>