Apply plug-in specific shortcut attributes only when launching shorcuts.
Since shortcuts do not set the "provider" attribute in their respective
configurations, a provider id is needed to get a shortcut and apply
plug-in specific attributes. Configurations that have the "provider"
attribute set are not shortcuts, and therefore do no need plug-in
specific attributes to be applied.

Change-Id: I5bcbe5a35bb7f245eac6be150ba643274744b43b
Reviewed-on: https://git.eclipse.org/r/7642
Tested-by: Hudson CI
Reviewed-by: Roland Grunberg <rgrunber@redhat.com>
IP-Clean: Roland Grunberg <rgrunber@redhat.com>
Tested-by: Roland Grunberg <rgrunber@redhat.com>