Change how runStap is passed the stap arguments
Currently runStap assumes that the only two commands used with systemtap
to parse possible probe points and functions are stap -p1 -v -e 'script'
and stap -u -p2 -e 'script'.  This might not always be the case and it
would be better to simply pass the full arguments themselves as a
parameter.  The functionality already exists to add the options passed
as a parameter, so we just use that for all of them.  Due to the level
no longer being appended in runStap, the parameter was removed from the
call.  In order to avoid breaking API the original runStap method has
been marked @deprecated and the new runStap marked @since 1.2 .

Change-Id: I425d2b5fe0d77aabbbbde1c834dd9c8b89d32aab
Reviewed-on: https://git.eclipse.org/r/7704
Tested-by: Hudson CI
Reviewed-by: Roland Grunberg <rgrunber@redhat.com>
IP-Clean: Roland Grunberg <rgrunber@redhat.com>
Tested-by: Roland Grunberg <rgrunber@redhat.com>