Fix BZ #389293 by making launch shortcut behvaiour consistent.
When launching a shortcut using a provider type (eg. memory, time,
snapshot), the user will be prompted to select any launch configuration
that may apply. The user should also be given the option to select a
launch configuration corresponding to the provider listed in the
preference page.

Change-Id: I6d335490df830fe3e093e0300ca2ff85db5ca928
Reviewed-on: https://git.eclipse.org/r/7755
Tested-by: Hudson CI
Reviewed-by: Camilo Bernal <cabernal@redhat.com>
IP-Clean: Camilo Bernal <cabernal@redhat.com>
Tested-by: Camilo Bernal <cabernal@redhat.com>