Make sure a launch provider exists for a plug-in when using a plug-in id.
If a plug-in id is acquired from the preferences, it's possible that
the plug-in corresponding to that id is not present. To prevent that,
a simple check is done, given the plug-in id, to make sure the launch
provider exists for the corresponding plug-in.

Change-Id: I51b880a7a2e9aa3127604d1126411992d164de99
Reviewed-on: https://git.eclipse.org/r/7854
Tested-by: Hudson CI
Reviewed-by: Roland Grunberg <rgrunber@redhat.com>
IP-Clean: Roland Grunberg <rgrunber@redhat.com>
Tested-by: Roland Grunberg <rgrunber@redhat.com>