Update project settings for CTF/LTTng/TMF
Remove the automatic line wrapping and unwrapping for code
lines. It had the tendency to absolutely destroy the formatting
if used on a complete file. The auto-wrapping of comments
remains at 80 columns.

Also enabled Javadoc warnings for protected fields and methods.
Since protected members in exported packages are part of the API,
those should be Javadoc-umented too.

Change-Id: Iab722d674e1f209f1e2b937ead9f058b9b09cab1
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/8473
Tested-by: Hudson CI
Reviewed-by: Francois Chouinard <fchouinard@gmail.com>
IP-Clean: Francois Chouinard <fchouinard@gmail.com>
Tested-by: Francois Chouinard <fchouinard@gmail.com>