Support for multiple events for same timestamp at checkpoint
If a trace has multiple events with the same timestamps before
a checkpoint and at the checkpoint, the seek to this timestamp
used to return the event at the checkpoint. The implementation
was changed so that the first event in the trace with this
timestamp is returned.

Also the compareTo() method of TmfCheckpoint has been updated
to consider the checkpoint location if the timestamp to search
is equal to the timestamp of a checkpoint. This ensure that
the first checkpoint is found of multiple checkpoints with same
timestamp

Change-Id: I73fa8f19d8a2dcc45da10bd44eaaa4e402f4a0dd
Reviewed-on: https://git.eclipse.org/r/8758
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Hudson CI
IP-Clean: Bernd Hufmann <bhufmann@gmail.com>
Tested-by: Bernd Hufmann <bhufmann@gmail.com>
Reviewed-by: Bernd Hufmann <bhufmann@gmail.com>