Make lower priority requests pre-emptible
The current request priority handling is based on splitting lower
priority requests into chunks of a fixed size that don't take too long
to process. If a higher priority request is issued (e.g. as the result
of a user action), it is simply put in front of the request queue i.e.
before the next low-priority chunk.

This patch modifies the TmfRequestExecutor to use 2 queues (one for each
priority) and implements a pre-emption scheme to make higher priority
requests run immediately.

This is the first step in a more comprehensive overhaul of the TMF
Request Model.


Change-Id: I6413cbbd69985d88b3fd5b5375a0b7ec59104682
Signed-off-by: Francois Chouinard <fchouinard@gmail.com>
Reviewed-on: https://git.eclipse.org/r/8983
Tested-by: Hudson CI