StapNewWizard fixed to work in workspace.
New wizards are better creating stuff in the workspace and not in random
place in the filesystem. This fixes a problem where if you had some
project selected the wizard was failing because of different roots -
system vs. workspace. 
It's a common thing for wizards to create stuff in the workspace after
all.

Change-Id: I72458f35765a5c6f272833116fabdf41e58baeea
Reviewed-on: https://git.eclipse.org/r/10067
Tested-by: Hudson CI
Reviewed-by: Sami Wagiaalla <swagiaal@redhat.com>
IP-Clean: Sami Wagiaalla <swagiaal@redhat.com>
Tested-by: Sami Wagiaalla <swagiaal@redhat.com>