ctf: Avoid re-downloading the test traces
Do not download the test traces if they are already present
on the file system (skipexisting="true").

The extraction was also moved to a conditional task, so that if
the traces cannot be downloaded for some reason, at least it
doesn't fail the complete build. Some unit tests will probably fail,
but that's the lesser of the two evils.

References bug 400459.

Change-Id: I69d7cb1525005cf6426f7aa42c45ecbb4c92cdff
Reviewed-on: https://git.eclipse.org/r/10286
Tested-by: Hudson CI
Reviewed-by: Alexander Kurtakov <akurtako@redhat.com>
IP-Clean: Alexander Kurtakov <akurtako@redhat.com>
Tested-by: Alexander Kurtakov <akurtako@redhat.com>