ctf: Allow the test traces to be used by other plugins
Centralize the management of the CTF test traces by moving the
"TestParams" file to a separate, exported package.

This required fiddling with the plugin type and adding an
activator, to mimic the behavior of "tmf.core.tests" (which
is also a test plugin that exports packages).

Change-Id: Ieeb32eda8099d797999a5803dcdc2b4f8c77a51b
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/10353
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bhufmann@gmail.com>
IP-Clean: Bernd Hufmann <bhufmann@gmail.com>
Tested-by: Bernd Hufmann <bhufmann@gmail.com>