Fix merged event iterator
The merged event iterator was sometimes creating split events with
negative duration when a zoomed event was longer than an unzoomed event
starting at the same time, and when the requested iterator range was
smaller than the zoomed events range. The algorithm for the iterator is
reworked. Also, comments.

Change-Id: I6ef83d3e3b9b25d716978cba906e9bef0985b5bd
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/10185
Reviewed-by: Bernd Hufmann <bhufmann@gmail.com>
IP-Clean: Bernd Hufmann <bhufmann@gmail.com>
Tested-by: Bernd Hufmann <bhufmann@gmail.com>
Tested-by: Hudson CI