Add transparent events to time graph
The time graph presentation provider interface is updated to define two
new constants for the state table index. When TRANSPARENT value is
returned, only the event border will be drawn, the event body will
remain transparent (empty) to allow the pixel to take the color of the
next visible event on the same pixel. The INVISIBLE value is defined,
although the functionality was previously supported. When INVISIBLE
value is returned, the event is not drawn at all and neither are its
borders.

The Time Chart view is updated to use the INVISIBLE value to hide
filtered events.

Change-Id: Ide68c4237db60824094f1f16739341cbf5e69860
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/10186
Reviewed-by: Bernd Hufmann <bhufmann@gmail.com>
IP-Clean: Bernd Hufmann <bhufmann@gmail.com>
Tested-by: Bernd Hufmann <bhufmann@gmail.com>
Tested-by: Hudson CI