Fix buffering issues for the Perf Source Disassembly View.
When running perf stat, the standard output stream contains the
program's output which can be of any size so it should be read first.

When running perf source disassembly the standard output stream contains
the relevant data and since it can be rather large in size, this stream
must be consumed first.

Change-Id: I4388a6275d57a33b5ac813eca5b52fb57ae8b8c8
Reviewed-on: https://git.eclipse.org/r/11405
Tested-by: Hudson CI
Reviewed-by: Camilo Bernal <cabernal@redhat.com>
IP-Clean: Camilo Bernal <cabernal@redhat.com>
Tested-by: Camilo Bernal <cabernal@redhat.com>