TMF: Add Variant field class to CtfTmfEventField
A variant field class was necessary to not lose the name of the variant
field.  The output of a variant field type is now identical to that of
babeltrace.

Modified some comments and added the getFormattedValue to CTFStructField
to correctly handle empty structs.

Removed overriden toString functions that are not overridden anymore.

Change-Id: Iccffc6a2375df54945ff41614628760afcdf147f
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/11550
Tested-by: Hudson CI
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
IP-Clean: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
IP-Clean: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>