tmf: Eliminate lesser constructors in TmfTrace
It's becoming very hard to figure out which exact constructor
is being called, and the resulting program flow. Remove the
intermediate constructors in TmfTrace and redirect implementations
to the full one.

Improved the Javadoc at the same time, to mention which default
values can be used.

Change-Id: Id0d03a976ce6d84e680021bf1634fa15d03942f3
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/11913
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
IP-Clean: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>