lttng: add support for kernel-only in tracer control (bug406120)
This commit allows the user to connect to a remote LTTng node 
where UST is disabled and only the kernel tracer is available. 
For LTTng tools 2.0.x, the command lttng list -u causes a 
segmentation fault on the remote side. This segmentation fault
makes it impossible to use the kernel-only mode. This bug is fixed
in later versions of LTTng tools starting from 2.1.0.

Change-Id: I2862ccce68c575706722eea89fdd4b450797d5d6
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/12139
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
IP-Clean: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Hudson CI
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>