tmf: Re-add null annotations to tmf.core.statevalue
Revert the revert in 9885a77c775e59e8ad72d5cc4495f4d1fea1c964.

Current master doesn't compile with Eclipse < 3.8 anymore, so since
this is our new oldest-supported version, we can put the null
annotations back in (those also require >= 3.8 to even compile).

Since the package is only needed at compilation and not at runtime,
it should not be added to the dependencies in MANIFEST.MF, but to
the build.properties file. As indicated in
http://help.eclipse.org/kepler/index.jsp?topic=%2Forg.eclipse.jdt.doc.user%2Ftasks%2Ftask-using_null_annotations.htm

Change-Id: I96a8bfd8f43296cf4eee196d0e44641f2822153c
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/16527
Tested-by: Hudson CI
IP-Clean: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>