tmf : introducing the mipmap
Introduce the AbstractTmfMipmapStateProvider class which can be used to
modify state system attributes while creating a mipmap of the attribute
for one or more of the supported features (max, min, average).

Add queries for max, min and average of an attribute over a range to the
state system operation. If the attribute was not built using the mipmap
state provider, the queries still work but query the base attribute
intervals over the full range.

Change-Id: Idfde25a4cc622db35c5e9bb8a3f88bc89285a6b4
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/14143
Tested-by: Hudson CI