all-tests: Run with less dependencies.
Instead of relying on sdk.[ide|feature] run the tests with
platform.[ide|feature]. There shouldn't be any difference for the tests
as what the sdk adds is the source bundles which are not used by the
tests. This should have some minimal effect to reduce our already way
too long build time.

Change-Id: I57bbc8c5ccb6e0a4575224134e184e13ae581b8f