tmf: Move default tracing project name to common constants
Currently, the default tracing project name "Tracing" is defined
in the RCP only. It makes sense to also create a default project in
other cases such as when importing a trace with no projects in the
workspace so the name needs to be accessible to other plug-ins.

Change-Id: I8739e651217f8b369a37d8b4784df5d9f8110e18
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/17944
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
IP-Clean: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>