tmf: Move the trace statistics to the analysis framework
By itself, this patch causes a bug where the Statistics view will not be
populated unless it is re-opened *after* the trace has been loaded.

The patch following this one fixes this (and unfortunately, that one
depends on changes from this one, so it wasn't trivial to re-order them).

Change-Id: I8b6c59d70e84fdc5971665e1e78d096e3abe30db
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/19412
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
IP-Clean: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Hudson CI