RPM: createrepo form editor
Add form editor with 3 pages: page to manage RPMs to create a repository
for, page to manage the repo metadata or customize the command
(could--and probably should--be put into a preference page), and lastly
a page to edit the .repo file.

This patch lacks most functionality and just contains what the pages
might look like. The most the repository page could do is to fill the
tree with RPMs found in the content folder.

Change-Id: I083ad8de7ba7af5e61d6d51bc857c7d650a29876
Signed-off-by: Neil Guzman <nguzman@redhat.com>
Reviewed-on: https://git.eclipse.org/r/18403
Reviewed-by: Alexander Kurtakov <akurtako@redhat.com>
IP-Clean: Alexander Kurtakov <akurtako@redhat.com>
Tested-by: Alexander Kurtakov <akurtako@redhat.com>