ctf: Update paths in the CTF-Testsuite tests
The upstream CTF-Testsuite was updated, let's follow 'suite'.

This also changes the test mechanisms to use parametrized tests. This
way, we get one JUnit test per test-trace being run, so we'll know the
exact amount of tests that are failing (instead of bailing out as soon
as we get one failure).

Those tests are disabled by default to avoid build failures for now,
but they should be fixed/reenabled as soon as possible.

Change-Id: Iadcd7c2d4352301cf60c132e0f2d8dc80f75a688
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/18479
Tested-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
IP-Clean: Matthew Khouzam <matthew.khouzam@ericsson.com>