tmf: Cleanup TmfStateValue to remove 1-level of indirection.
This is a refactoring to diminush the memory usage of the
StateSystem.

In the current implementation, each TmfStateValue has a
"box" for an other object (i.e. Integer, Double, String, ...).
We propose to remove this indirection and use the raw type inside
the corresponding type.

As an example, IntegerStateValue now contains an "int" instead
of an "Integer".

Change-Id: Iaa75172be42174ebd6dc7d9bdd434631077c0146
Signed-off-by: Etienne Bergeron <etienne.bergeron@gmail.com>
Reviewed-on: https://git.eclipse.org/r/19027
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
IP-Clean: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
IP-Clean: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>